requirements_updated.txt
flask
requests
psycopg2-binary
python-dotenv
gunicorn
waitress
Pillow
torch
torchvision
torchaudio
spacy